Fix args bug
Sorry, i've been sloppy in keeping track of my commits. i deployed the changes to tethys2 and noticed i didn't have this fix in the container when i couldn't get any results. I had committed it on my computer but never pushed it so it worked locally but not when i ran it on tethys2.

but on the bright side, it works on tethys2 now. for example: https://tethys2.byu.edu/localsptapi/api/GetReachID/?lat=10&lon=10